extracted createPageContentsContext to a seperate file
